Moved DijkstraMap wrapper closer to nativescript.gdns
